Vec2 work

+++ external/glm/core/intrinsic_common.inl
@@ -8,10 +8,10 @@
 /// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 /// copies of the Software, and to permit persons to whom the Software is
 /// furnished to do so, subject to the following conditions:
-/// 
+///
 /// The above copyright notice and this permission notice shall be included in
 /// all copies or substantial portions of the Software.
-/// 
+///
 /// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 /// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 /// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -41,7 +41,7 @@ namespace detail{
 	   {
 		  const unsigned int mantissa:23, exp:8, sign:1;
 	   };
-	   
+
 	   ieee754_QNAN() : f(0.0)/*, mantissa(0x7FFFFF), exp(0xFF), sign(0x0)*/ {}
 	};
 
@@ -62,7 +62,7 @@ namespace detail{
 	static const ieee754_QNAN absMask;
 	static const __m128 GLM_VAR_USED abs4Mask = _mm_set_ps1(absMask.f);
 
-	static const __m128 GLM_VAR_USED _epi32_sign_mask = _mm_castsi128_ps(_mm_set1_epi32(static_cast<int>(0x80000000)));
+	static const __m128 GLM_VAR_USED _epi32_sign_mask = _mm_castsi128_ps(_mm_set1_epi32(static_cast<int>((int)(int)0x80000000)));
         //static const __m128 GLM_VAR_USED _epi32_inv_sign_mask = _mm_castsi128_ps(_mm_set1_epi32(0x7FFFFFFF));
         //static const __m128 GLM_VAR_USED _epi32_mant_mask = _mm_castsi128_ps(_mm_set1_epi32(0x7F800000));
         //static const __m128 GLM_VAR_USED _epi32_inv_mant_mask = _mm_castsi128_ps(_mm_set1_epi32(0x807FFFFF));
@@ -154,7 +154,7 @@ namespace detail{
 GLM_FUNC_QUALIFIER __m128 sse_abs_ps(__m128 x)
 {
 	return _mm_and_ps(glm::detail::abs4Mask, x);
-} 
+}
 
 GLM_FUNC_QUALIFIER __m128 sse_sgn_ps(__m128 x)
 {
