+++ external/glm/gtc/ulp.inl
@@ -8,10 +8,10 @@
 /// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 /// copies of the Software, and to permit persons to whom the Software is
 /// furnished to do so, subject to the following conditions:
-/// 
+///
 /// The above copyright notice and this permission notice shall be included in
 /// all copies or substantial portions of the Software.
-/// 
+///
 /// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 /// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 /// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -101,16 +101,16 @@ namespace detail
 		ix = hx&0x7fffffff;             // |x|
 		iy = hy&0x7fffffff;             // |y|
 
-		if((ix>0x7f800000) ||   // x is nan 
-			(iy>0x7f800000))     // y is nan 
+		if((ix>0x7f800000) ||   // x is nan
+			(iy>0x7f800000))     // y is nan
 			return x+y;
 		if(x==y) return y;              // x=y, return y
 		if(ix==0) {                             // x == 0
-			GLM_SET_FLOAT_WORD(x,(hy&0x80000000)|1);// return +-minsubnormal
+			GLM_SET_FLOAT_WORD(x,(hy&(int)(int)0x80000000)|1);// return +-minsubnormal
 			t = x*x;
 			if(t==x) return t; else return x;   // raise underflow flag
 		}
-		if(hx>=0) {                             // x > 0 
+		if(hx>=0) {                             // x > 0
 			if(hx>hy) {                         // x > y, x -= ulp
 				hx -= 1;
 			} else {                            // x < y, x += ulp
@@ -144,27 +144,27 @@ namespace detail
 
 		GLM_EXTRACT_WORDS(hx, lx, x);
 		GLM_EXTRACT_WORDS(hy, ly, y);
-		ix = hx & 0x7fffffff;             // |x| 
-		iy = hy & 0x7fffffff;             // |y| 
+		ix = hx & 0x7fffffff;             // |x|
+		iy = hy & 0x7fffffff;             // |y|
 
 		if(((ix>=0x7ff00000)&&((ix-0x7ff00000)|lx)!=0) ||   // x is nan
 			((iy>=0x7ff00000)&&((iy-0x7ff00000)|ly)!=0))     // y is nan
 			return x+y;
 		if(x==y) return y;              // x=y, return y
-		if((ix|lx)==0) {                        // x == 0 
-			GLM_INSERT_WORDS(x, hy & 0x80000000, 1);    // return +-minsubnormal
+		if((ix|lx)==0) {                        // x == 0
+			GLM_INSERT_WORDS(x, hy & (int)(int)0x80000000, 1);    // return +-minsubnormal
 			t = x*x;
-			if(t==x) return t; else return x;   // raise underflow flag 
+			if(t==x) return t; else return x;   // raise underflow flag
 		}
-		if(hx>=0) {                             // x > 0 
-			if(hx>hy||((hx==hy)&&(lx>ly))) {    // x > y, x -= ulp 
+		if(hx>=0) {                             // x > 0
+			if(hx>hy||((hx==hy)&&(lx>ly))) {    // x > y, x -= ulp
 				if(lx==0) hx -= 1;
 				lx -= 1;
 			} else {                            // x < y, x += ulp
 				lx += 1;
 				if(lx==0) hx += 1;
 			}
-		} else {                                // x < 0 
+		} else {                                // x < 0
 			if(hy>=0||hx>hy||((hx==hy)&&(lx>ly))){// x < y, x -= ulp
 				if(lx==0) hx -= 1;
 				lx -= 1;
